Implement persistent database connections option
Persistence attribute has been added to the Shard object, which controls whether or not the database connection should be maintained. This attribute is then utilized in the CDO object to set the appropriate connection option. As a result, it is now possible to create persistent database connections through Shard and CDO classes. The version documentation for both classes has been updated accordingly.